@nalimilan 2021-04-11
> by despecializing we incur a cost of having a loop over threads that is not specialized and will use dynamic dispatch but I assume that this loop is small (up to number of available threads). Is this correct @nalimilan?

Sounds reasonable, but can you check that performance isn't affected for large tables?