@bkamins 2021-04-11
> Sounds reasonable, but can you check that performance isn't affected for large tables?

Yes. I will do some timings (these always take time) and report here (probably today in the evening or tomorrow).