@bkamins 2021-04-11
This is super strange.

`main` timing of grouping tests run twice are:
```
417.883939 seconds (1.09 G allocations: 61.362 GiB, 5.79% gc time, 0.00% compilation time)
171.446700 seconds (574.79 M allocations: 31.722 GiB, 5.93% gc time, 0.10% compilation time)
```

with this PR wrapping `Val` in `Ref` I get:
```
394.539858 seconds (1.06 G allocations: 59.078 GiB, 5.72% gc time, 0.00% compilation time)
171.479467 seconds (575.52 M allocations: 31.758 GiB, 5.65% gc time, 0.10% compilation time)
```

now if I pass `Bool` and only switch it to `Ref` at the end I get:
```
398.629919 seconds (1.08 G allocations: 60.430 GiB, 5.81% gc time, 0.00% compilation time)
177.364990 seconds (590.24 M allocations: 32.977 GiB, 5.87% gc time, 0.10% compilation time)
```

which is quite puzzling.

Therefore I decided to go for a cleaner design and introduce types that signal if we have a single or multi column aggregation. With them the timing is:
```
393.826773 seconds (1.06 G allocations: 59.289 GiB, 5.70% gc time, 0.00% compilation time)
174.152069 seconds (578.06 M allocations: 31.944 GiB, 5.83% gc time, 0.10% compilation time)
```
so this is comparable to passing around `Val` wrapped in `Ref` but I would go for it as the design is much more clean I think with it. This is what I have pushed.

Now I am benchmarking the performance on larger split-apply-combine operations.