@bkamins 2021-04-11
The performance benchmarks show that this PR does not regress against `main` and in general we have a strict improvement over 0.22.7 release. Could you just quickly have a look at 1 vs 4 threads timings to confirm that the cases where threads speed up things are as as expected (this is my understanding as I understand currently combining rows is only threaded, but please confirm).

Here are the details:

```
# -------- setup code
julia> using DataFrames, BenchmarkTools
julia> df = DataFrame(a=repeat(1:10^6, 10), b=1:10^7);
julia> gdf = groupby(df, :a);
julia> f1(x) = sum(x);
julia> f2(x) = [sum(x), 1];
julia> f3(x) = (x=sum(x), y=1);
julia> f4(x) = (x=[sum(x), 1], y=[1, 1]);

# -------- PR 1 thread
julia> @btime combine($gdf, :b => sum);
  21.486 ms (210 allocations: 23.86 MiB)
julia> @btime combine($gdf, :b => f1);
  232.539 ms (1000289 allocations: 175.49 MiB)
julia> @btime combine($gdf, :b => f2);
  951.362 ms (15999252 allocations: 644.35 MiB)
julia> @btime combine($gdf, :b => f3 => AsTable);
  287.552 ms (1000337 allocations: 190.75 MiB)
julia> @btime combine($gdf, :b => f4 => AsTable);
  1.050 s (16999331 allocations: 798.68 MiB)

# -------- PR 4 threads
julia> @btime combine($gdf, :b => sum);
  21.261 ms (212 allocations: 23.86 MiB)
julia> @btime combine($gdf, :b => f1);
  98.041 ms (1000334 allocations: 175.49 MiB)
julia> @btime combine($gdf, :b => f2);
  937.994 ms (15999254 allocations: 644.35 MiB)
julia> @btime combine($gdf, :b => f3 => AsTable);
  112.330 ms (1000383 allocations: 190.76 MiB)
julia> @btime combine($gdf, :b => f4 => AsTable);
  1.069 s (16999334 allocations: 798.68 MiB)

# -------- main 1 thread
julia> @btime combine($gdf, :b => sum);
  21.530 ms (208 allocations: 23.86 MiB)
julia> @btime combine($gdf, :b => f1);
  289.125 ms (1000267 allocations: 175.49 MiB)
julia> @btime combine($gdf, :b => f2);
  925.847 ms (15999252 allocations: 644.35 MiB)
julia> @btime combine($gdf, :b => f3 => AsTable);
  292.836 ms (1000315 allocations: 190.75 MiB)
julia> @btime combine($gdf, :b => f4 => AsTable);
  1.058 s (16999331 allocations: 798.68 MiB)

# -------- main 4 threads
julia> @btime combine($gdf, :b => sum);
  21.382 ms (210 allocations: 23.86 MiB)
julia> @btime combine($gdf, :b => f1);
  96.502 ms (1000318 allocations: 175.49 MiB)
julia> @btime combine($gdf, :b => f2);
  951.696 ms (15999254 allocations: 644.35 MiB)
julia> @btime combine($gdf, :b => f3 => AsTable);
  119.321 ms (1000367 allocations: 190.76 MiB)
julia> @btime combine($gdf, :b => f4 => AsTable);
  1.093 s (16999334 allocations: 798.68 MiB)

# -------- 0.22.7 1 thread
julia> @btime combine($gdf, :b => sum);
  21.310 ms (169 allocations: 23.85 MiB)
julia> @btime combine($gdf, :b => f1);
  269.108 ms (9999177 allocations: 312.80 MiB)
julia> @btime combine($gdf, :b => f2);
  1.081 s (20999215 allocations: 842.71 MiB)
julia> @btime combine($gdf, :b => f3 => AsTable);
  388.781 ms (9999193 allocations: 404.36 MiB)
julia> @btime combine($gdf, :b => f4 => AsTable);
  1.268 s (22999248 allocations: 1.02 GiB)

# -------- 0.22.7 4 threads
julia> @btime combine($gdf, :b => sum);
  21.947 ms (169 allocations: 23.85 MiB)
julia> @btime combine($gdf, :b => f1);
  270.780 ms (9999177 allocations: 312.80 MiB)
julia> @btime combine($gdf, :b => f2);
  1.186 s (20999215 allocations: 842.71 MiB)
julia> @btime combine($gdf, :b => f3 => AsTable);
  385.580 ms (9999193 allocations: 404.36 MiB)
julia> @btime combine($gdf, :b => f4 => AsTable);
  1.371 s (22999248 allocations: 1.02 GiB)
```
in summary - this PR looks good.